+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## v1.3.0 (2023-11-XX)
 
+### Bug Fixes
+
+- Add missing `nav` CSS Part to `a-tab-group` element.
+
+### Other Changes
+
 - Add default display values to host elements (`a-tab`, `a-tab-panel`).
 - Add `contain: content` to host elements (`a-tab-group`, `a-tab`, `a-tab-panel`) to prevent layout thrashing.
 - Provide better ergonomics for `a-tab` and `a-tab-panel` by automatically setting the `slot` attribute to `tab` and `panel` respectively, if not provided.
+- Add `nav--scrollable` CSS Part to navigation container when tabs are scrollable, to allow for more specific styling.
 - Generate types declarations for the components.
 
 ## v1.2.0 (2023-10-13)